python/apps/taiga/requirements/prod.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements/prod.in
#
aiopg==1.3.4
    # via procrastinate
aiosmtplib==1.1.6
    # via fastapi-mailman
anyio==3.6.1
    # via
    #   httpcore
    #   starlette
    #   watchfiles
asgiref==3.5.2
    # via django
async-timeout==4.0.2
    # via aiopg
asyncpg==0.25.0
    # via -r requirements/commons.in
attrs==22.1.0
    # via procrastinate
certifi==2022.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.0
    # via requests
click==8.1.3
    # via
    #   procrastinate
    #   typer
    #   uvicorn
croniter==1.3.5
    # via procrastinate
cryptography==37.0.4
    # via -r requirements/commons.in
cssselect2==0.6.0
    # via svglib
django==4.1
    # via
    #   -r requirements/commons.in
    #   django-stubs
    #   django-stubs-ext
    #   easy-thumbnails
django-stubs==1.12.0
    # via -r requirements/commons.in
django-stubs-ext==0.5.0
    # via django-stubs
dnspython==2.2.1
    # via
    #   email-validator
    #   fastapi-mailman
easy-thumbnails==2.8.3
    # via -r requirements/commons.in
email-validator==1.2.1
    # via
    #   fastapi-mailman
    #   pydantic
faker==13.15.1
    # via -r requirements/commons.in
fastapi==0.79.0
    # via
    #   -r requirements/commons.in
    #   fastapi-mailman
fastapi-mailman==0.0.4
    # via -r requirements/commons.in
h11==0.12.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.15.0
    # via httpx
httptools==0.4.0
    # via uvicorn
httpx==0.23.0
    # via -r requirements/commons.in
idna==3.3
    # via
    #   anyio
    #   email-validator
    #   requests
    #   rfc3986
jinja2==3.1.2
    # via
    #   -r requirements/commons.in
    #   fastapi-mailman
lxml==4.9.1
    # via svglib
markupsafe==2.1.1
    # via jinja2
mypy==0.971
    # via django-stubs
mypy-extensions==0.4.3
    # via mypy
pillow==9.2.0
    # via
    #   easy-thumbnails
    #   reportlab
procrastinate==0.25.0
    # via -r requirements/commons.in
psycopg2==2.9.3
    # via -r requirements/commons.in
psycopg2-binary==2.9.3
    # via
    #   aiopg
    #   procrastinate
pycparser==2.21
    # via cffi
pydantic[dotenv,email]==1.9.1
    # via
    #   -r requirements/commons.in
    #   fastapi
    #   fastapi-mailman
pyhumps==3.7.2
    # via -r requirements/commons.in
pyjwt==2.4.0
    # via -r requirements/commons.in
python-dateutil==2.8.2
    # via
    #   croniter
    #   faker
    #   procrastinate
python-dotenv==0.20.0
    # via
    #   pydantic
    #   uvicorn
python-json-logger==2.0.4
    # via -r requirements/commons.in
python-multipart==0.0.5
    # via -r requirements/commons.in
python-slugify==6.1.2
    # via -r requirements/commons.in
pyyaml==6.0
    # via uvicorn
reportlab==3.6.11
    # via svglib
requests==2.28.1
    # via -r requirements/commons.in
rfc3986[idna2008]==1.5.0
    # via httpx
six==1.16.0
    # via
    #   python-dateutil
    #   python-multipart
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlparse==0.4.2
    # via django
starlette==0.19.1
    # via fastapi
svglib==1.4.1
    # via -r requirements/commons.in
text-unidecode==1.3
    # via python-slugify
tinycss2==1.1.1
    # via
    #   cssselect2
    #   svglib
tomli==2.0.1
    # via
    #   django-stubs
    #   mypy
typer==0.6.1
    # via -r requirements/commons.in
types-pytz==2022.1.2
    # via django-stubs
types-pyyaml==6.0.11
    # via django-stubs
typing-extensions==4.3.0
    # via
    #   django-stubs
    #   django-stubs-ext
    #   mypy
    #   pydantic
urllib3==1.26.11
    # via requests
uvicorn[standard]==0.18.2
    # via -r requirements/commons.in
uvloop==0.16.0
    # via uvicorn
watchfiles==0.16.1
    # via uvicorn
webencodings==0.5.1
    # via
    #   cssselect2
    #   tinycss2
websockets==10.3
    # via uvicorn
